user must be logged in to bid
-replace redirect after action in SessionsController
-redirect back to item page when clicking "Login or Create Account to
Bid"
-add a div id to the new_bid form
-combine some of the bid specs into one
-add the two contexts "user not logged in" and "user is logged in" to
the bid spec file
